Updated Profile Page Functionality
The Profile page now displays the current users profile, and allows them to upload a new pfp. Any new user will automatically have the default picture pfp. I also added the bio field to the users table, and finished the functionality to save and display the users bio in the profile page.