Fixing Group Chat Creation and CSS
Fixed Group Chat Creation. New group chat items will now display a defult clear image as the last sent message. They will also display "New Group Chat" as the last sent message, and "New" as the timestamp.

I also fixed a lot of the css issues on the chat page. I moved the groupchat items to the left side of the screen, similar to how the frinds are pushed all the way to the right. I centered the chat area, and made it so long messages automatically wrap. Group Chat Items will now only display a few characters before the fade to ... This is to prevent entire paragraphs from being loaded in the group chat items area,